fix(cache): address shared index review feedback
