+++ crates/rginx-http/src/cache/mod.rs
@@ -41,13 +41,10 @@ use request::{cache_request_bypass, render_cache_key};
 use runtime::PurgeSelector;
 pub(crate) use runtime::with_cache_status;
 use runtime::{build_conditional_headers, remove_cache_files_if_unindexed};
-use shared::{
-    SharedIndexStore, bootstrap_shared_index, remove_zone_shared_index_entry,
-    sync_zone_shared_index_if_needed,
-};
+use shared::{SharedIndexStore, bootstrap_shared_index, sync_zone_shared_index_if_needed};
 use store::{
     CacheStoreError, cleanup_inactive_entries_in_zone, lock_index, purge_zone_entries,
-    refresh_not_modified_response, remove_index_entry, store_response,
+    refresh_not_modified_response, remove_zone_index_entry, store_response,
 };
 
 const CACHE_STATUS_HEADER: &str = "x-cache";
